Use projected token source to dynamically refresh SA token
Watching the SA token file is pointless, as it will never change,
because token secret rotation means a new secret gets created. This new
secret will never get used in an old Pod though.

The way to go for this are [projected service acccount token volumes][0]
and setting BeearerTokenFile on the restcfg, this will make client-go
periodically reload it.

[0]: https://kubernetes.io/docs/tasks/configure-pod-container/configure-service-account/#service-account-token-volume-projection